+++ src/common_restrictions/coupled_dofs_restriction.jl
@@ -33,6 +33,7 @@ function assemble!(R::CoupledDofsRestriction, sol, SC; kwargs...)
 
     R.parameters[:matrix] = B
     R.parameters[:rhs] = Zeros(length(unique_cols))
+    R.parameters[:fixed_dofs] = unique_cols
 
     return nothing
 end